requirements.txt
appnope (0.1.0)
backports.shutil-get-terminal-size (1.0.0)
beautifulsoup4 (4.5.1)
decorator (4.0.10)
Django (1.10.2)
enum34 (1.1.6)
get (0.0.0)
image (1.5.5)
ipdb (0.10.1)
ipython (5.1.0)
ipython-genutils (0.1.0)
pathlib2 (2.1.0)
pexpect (4.2.1)
pickleshare (0.7.4)
Pillow (3.4.2)
pip (8.1.2)
post (0.0.0)
prompt-toolkit (1.0.8)
ptyprocess (0.5.1)
public (0.0.0)
Pygments (2.1.3)
query-string (0.0.0)
requests (2.11.1)
setupfiles (0.0.13)
setuptools (28.6.0)
simplegeneric (0.8.1)
six (1.10.0)
traitlets (4.3.1)
wcwidth (0.1.7)
wheel (0.30.0a0)
